+++ src/system/mod.rs
@@ -2,4 +2,3 @@
 
 pub(crate) mod environment;
 pub(crate) mod fs;
-pub(crate) mod steam;